# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZenMon API Documentation",
        "description": "Monitoring Application - Alternative to Zabbix/Nagios",
        "contact": {
            "email": "admin@zenmon.local"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "ZenMon API Server"
        }
    ],
    "paths": {
...
        "/api/host-configurations": {
            "get": {
                "tags": [
                    "Host Configurations"
                ],
                "summary": "Get list of host configurations (UC24)",
                "description": "Returns list of host configurations with filtering options",
                "operationId": "getHostConfigurationsList",
                "parameters": [
                    {
                        "name": "host_id",
                        "in": "query",
                        "description": "Filter by host ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/HostConfiguration"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Host Configurations"
                ],
                "summary": "Create new host configuration (UC24)",
                "description": "Create new monitoring configuration for host",
                "operationId": "createHostConfiguration",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "host_id"
                                ],
                                "properties": {
                                    "host_id": {
                                        "description": "Host ID",
                                        "type": "integer"
                                    },
                                    "data_collection_interval": {
                                        "description": "Interval in seconds (30-3600)",
                                        "type": "integer",
                                        "default": 120
                                    },
                                    "enable_cpu_monitoring": {
                                        "type": "boolean",
                                        "default": true
                                    },
                                    "enable_ram_monitoring": {
                                        "type": "boolean",
                                        "default": true
                                    },
                                    "enable_disk_monitoring": {
                                        "type": "boolean",
                                        "default": true
                                    },
                                    "enable_network_monitoring": {
                                        "type": "boolean",
                                        "default": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Configuration created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/HostConfiguration"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "409": {
                        "description": "Configuration already exists"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/host-configurations/{hostConfiguration}": {
            "get": {
                "tags": [
                    "Host Configurations"
                ],
                "summary": "Get specific host configuration (UC24)",
                "description": "Get configuration details for specific configuration",
                "operationId": "showHostConfiguration",
                "parameters": [
                    {
                        "name": "hostConfiguration",
                        "in": "path",
                        "description": "Configuration ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/HostConfiguration"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Configuration not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Host Configurations"
                ],
                "summary": "Update host configuration (UC24)",
                "description": "Update monitoring configuration for host",
                "operationId": "updateHostConfiguration",
                "parameters": [
                    {
                        "name": "hostConfiguration",
                        "in": "path",
                        "description": "Configuration ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data_collection_interval": {
                                        "description": "Interval in seconds (30-3600)",
                                        "type": "integer"
                                    },
                                    "enable_cpu_monitoring": {
                                        "type": "boolean"
                                    },
                                    "enable_ram_monitoring": {
                                        "type": "boolean"
                                    },
                                    "enable_disk_monitoring": {
                                        "type": "boolean"
                                    },
                                    "enable_network_monitoring": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Configuration updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/HostConfiguration"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Configuration not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Host Configurations"
                ],
                "summary": "Delete host configuration (UC24)",
                "description": "Delete monitoring configuration (resets to defaults)",
                "operationId": "deleteHostConfiguration",
                "parameters": [
                    {
                        "name": "hostConfiguration",
                        "in": "path",
                        "description": "Configuration ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuration deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Configuration not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HostConfiguration": {
                "title": "HostConfiguration",
                "description": "Host monitoring configuration",
                "properties": {
                    "host_configuration_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "host_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "collection_interval": {
                        "description": "Interval in seconds",
                        "type": "integer",
                        "example": 120
                    },
                    "auto_discovery": {
                        "type": "boolean",
                        "example": true
                    },
                    "monitoring_settings": {
                        "type": "object",
                        "example": {
                            "cpu_enabled": true,
                            "ram_enabled": true
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "host": {
                        "properties": {
                            "host_id": {
                                "type": "integer"
                            },
                            "host_name": {
                                "type": "string"
                            },
                            "ip_address": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}